refactor(oma-pm): move `PackageMatcher::search` outside
