+++ src/subcommand/search.rs
@@ -3,9 +3,9 @@ use std::borrow::Cow;
 use oma_console::{indicatif::ProgressBar, pager::Pager, pb::spinner_style};
 use oma_pm::{
     apt::{AptConfig, OmaApt, OmaAptArgs},
-    matches::{PackagesMatcher, SearchEngine},
+    matches::SearchEngine,
+    search::{IndiciumSearch, OmaSearch, SearchResult, StrSimSearch, TextSearch},
 };
-use oma_utils::dpkg::dpkg_arch;
 use tracing::warn;
 
 use crate::{error::OutputError, table::oma_display_with_normal_output};
@@ -20,18 +20,12 @@ pub fn execute(
     json: bool,
     another_apt_options: Vec<String>,
 ) -> Result<i32, OutputError> {
-    let arch = dpkg_arch(&sysroot)?;
-
     let oma_apt_args = OmaAptArgs::builder()
         .another_apt_options(another_apt_options)
         .sysroot(sysroot)
         .build();
 
     let apt = OmaApt::new(vec![], oma_apt_args, false, AptConfig::new())?;
-    let matcher = PackagesMatcher::builder()
-        .cache(&apt.cache)
-        .native_arch(&arch)
-        .build();
 
     let s = args.concat();
 
@@ -47,7 +41,8 @@ pub fn execute(
         None
     };
 
-    let res = matcher.search(
+    let res = search(
+        &apt,
         &s,
         match engine.as_str() {
             "indicium" => SearchEngine::Indicium(Box::new(|_| {})),
@@ -98,4 +93,20 @@ pub fn execute(
     })?;
 
     Ok(exit.into())
+}
+
+pub fn search(
+    apt: &OmaApt,
+    keyword: &str,
+    engine: SearchEngine,
+) -> Result<Vec<SearchResult>, OutputError> {
+    let searcher: Box<dyn OmaSearch> = match engine {
+        SearchEngine::Indicium(f) => Box::new(IndiciumSearch::new(&apt.cache, f)?),
+        SearchEngine::Strsim => Box::new(StrSimSearch::new(&apt.cache)),
+        SearchEngine::Text => Box::new(TextSearch::new(&apt.cache)),
+    };
+
+    let res = searcher.search(keyword)?;
+
+    Ok(res)
 }
